enable setting usage
This will allow us to save multiple discord instances, and allow each server to have their own leaderboards, and CB Results displayed if they so wish it.

It will also allow the enabling of rift codes, allowing codes to be sent to all designated "Code" channels through the approval of an admin.